requirements.txt
accelerate==0.28.0
bitsandbytes==0.43.1
diffusers==0.25.1
fvcore==0.1.5.post20221221
huggingface_hub==0.23.2
imageio==2.28.1
lmdb==1.4.1
lpips==0.1.4
matplotlib==3.7.1
numpy==1.24.3
# nvdiffrast==0.3.1
# nvdiffrast
omegaconf
# omegaconf==2.3.0
open3d==0.18.0
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.5.5.64
pandas==2.0.1
Pillow==9.5.0
PyMCubes==0.1.4
pytorch_lightning==2.1.2
rembg==2.0.57
setuptools==66.0.0
# tinycudann==1.7
# tinycudann
torch==2.0.1
torchmetrics==0.11.4
torchvision
tqdm==4.66.4
transformers==4.25.1
trimesh==4.3.2
typeguard==2.13.3
webdataset==0.2.86
xatlas==0.0.9
xformers==0.0.22
